feature#19 사이드바 백엔드 데이터 받는 중
